Add files via upload
Poprawiono funkcje wyznaczania trasy.
Brakujace funkcjionalnosci opisane w zalaczonym pliku.

+++ LCS_New_v1/SourceCode/tor.cpp
@@ -50,7 +50,7 @@ void Tor::setITrain()
     }
 }
 
-void Tor::setLen(int num)
+void Tor::setLen(double num)
 {
     Length = num;
 }
@@ -75,7 +75,7 @@ int Tor::getITrain()
     return IsTrain;
 }
 
-int Tor::getLen()
+double Tor::getLen()
 {
     return Length;
 }
